requirements.txt
streamlit==1.32.0
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
torch==2.2.1
transformers==4.38.2
plotly==5.19.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 